Big step towards making GiL a one time interface to Gecode
Problem creation, solver creation and solution request are now done exclusively in Gecode, you just need to call 3 lisp functions to use it from OpenMusic. However, solutions are not displayed in lisp as of now, but they are computed correctly and are printed in the log.txt file that is located at the root of the file system by default.